query-cost-optimizer/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "query-cost-optimizer"
version = "0.1.0"
description = "Cost & Performance Optimization Engine for BigQuery and Snowflake"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
keywords = ["bigquery", "snowflake", "sql", "optimization", "cost", "performance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Topic :: Database",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "pandas>=2.0",
    "sqlglot>=23.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "tabulate>=0.9",
    "google-cloud-bigquery>=3.0; extra == 'bigquery'",
    "snowflake-connector-python>=3.0; extra == 'snowflake'",
]

[project.optional-dependencies]
bigquery = ["google-cloud-bigquery>=3.0", "google-cloud-bigquery-storage>=2.0"]
snowflake = ["snowflake-connector-python>=3.0"]
all = [
    "google-cloud-bigquery>=3.0",
    "google-cloud-bigquery-storage>=2.0",
    "snowflake-connector-python>=3.0",
]
dev = ["pytest>=7.0", "pytest-cov>=4.0", "black>=23.0", "ruff>=0.1", "mypy>=1.0"]

[project.scripts]
qco = "query_cost_optimizer.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
